fix(lfm2): remove spurious GGUF conv-weight transpose, fix l_cache kernel-size bug, correct safetensors data offset
Three bugs fixed to get LFM2.5 350M inference working on both safetensors (BF16)
and GGUF (Q8_0) backends:

1. safetensors data offset: read_safetensor_bytes was using absolute file offset
   instead of data_region_start + relative offset. Reverted to correct formula.

2. l_cache kernel-size: ShortConv forward used shape.dims()[1] to get kernel
   size, but for 3D conv weight [hidden, 1, kernel] this returned 1 (in_channels)
   instead of 3 (kernel_size). Fixed by using .last() to always get kernel dim.

3. GGUF conv-weight transpose (ROOT CAUSE of gibberish): The GGUF converter
   stores conv weight data in [hidden, kernel] row-major order (verified
   byte-for-byte against safetensors source), but lfm2.rs assumed GGUF stored
   [kernel, hidden] and applied an explicit transpose that corrupted every
   shortconv layer. Removed the incorrect transpose; both paths now treat
   the flat data as canonical [hidden, kernel] layout.

Also includes: LFM2 model loader, GGUF tokenizer, sampler, streaming forward,
ROCm device/kernels, CLI run/train/verify commands, server, and garage
improvements from prior sessions.

+++ crates/grim-backend-rocm/src/kernels/fused_dequant_gemm.rs
@@ -66,12 +66,12 @@ extern "C" {
         int backup_codes_offset,
         int backup_scale_offset)
     {
-        const int idx = blockIdx.x * blockDim.x + threadIdx.x;
-        const int total = M * N;
+        const unsigned long long idx = (unsigned long long)blockIdx.x * blockDim.x + threadIdx.x;
+        const unsigned long long total = (unsigned long long)M * N;
         if (idx >= total) return;
 
-        const int row = idx / N;
-        const int col = idx % N;
+        const int row = (int)(idx / N);
+        const int col = (int)(idx % N);
 
         float scale = 1.0f;
         if (B_scales != nullptr) {
@@ -119,12 +119,12 @@ extern "C" {
         int backup_codes_offset,
         int backup_scale_offset)
     {
-        const int idx = blockIdx.x * blockDim.x + threadIdx.x;
-        const int total = M * K;
+        const unsigned long long idx = (unsigned long long)blockIdx.x * blockDim.x + threadIdx.x;
+        const unsigned long long total = (unsigned long long)M * K;
         if (idx >= total) return;
 
-        const int row = idx / K;
-        const int k = idx % K;
+        const int row = (int)(idx / K);
+        const int k = (int)(idx % K);
 
         float acc = 0.0f;
         for (int col = 0; col < N; ++col) {
